pyproject.toml
[tool.poetry]
name = "blaatur-backend"
version = "0.1.0"
description = "Backend for Blaatur web.app."
authors = ["Zylvian <33404765+Zylvian@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9.0"
Flask = "^1.1.2"
gunicorn = "^20.0.4"
Flask-Cors = "^3.0.10"
pytest = "^6.2.2"
requests = "^2.25.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
